refactor(runtime): split event-source adapters by source
Closes THE-2424.
